HttpCache: Network transaction and member of Writers is not mutual exclusive.
This CL removes the dcheck from HttpCache::Transaction::SetPriority that
presence of network_trans_ implies !InWriters(). For a range request, it
is possible that the transaction is in writers but is in the headers phase
to serve a subsequent range and network_trans_ is the network transaction
created for that range.

Bug: 785561
Change-Id: If0c25209e3c1eefd911a8f544834e3f6604001d9
Reviewed-on: https://chromium-review.googlesource.com/820072
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#523197}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c268c911c7def0c5431a39b612ca83e1a6f5e02a