Fix cachetool's dumping of stream 1.
GetStreamForKeyBuffer keeps incrementing offset in GrowableIOBuffer when
it collects things, so at the end data() actually points past the end
(like offset does), and one needs StartOfBuffer() to get to the beginning.

The other implementation of ReturnBuffer already did this.

With this, one can do something like:
 out/Default/cachetool ~/.cache/google-chrome-beta/Default/Cache/ simple get_stream https://www.chromium.org/ 1  > /tmp/file.txt
to see what's in the cache --- and in chromium.org's case, I needed zcat to view it, since it was gzipped.
(passing 0 instead of 1 would demarshal the HttpResponseInfo that goes with the body).

Change-Id: Id3b5608e2845f9404c2c3c21d53e7ce723f13320
Reviewed-on: https://chromium-review.googlesource.com/825102
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#523845}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b4c032d2d0928dcdde16f9a03d060f0769ee3ba4