Switch QUIC browser tests to make real requests to determine if QUIC is enabled.
This allows the test to work when the URLRequestContext is in another process.

This change also adds new certificates for testing QUIC, as the existing ones don't work with the
production CertVerifier (thanks to estark@ who diagnosed this and provided the fix). This also
includes a fix for QuicHttpResponseCache to work with URLs that have ports.

Bug: 793769
Change-Id: Ia03a18469df064258f7f5cde6b2578d6df137500
Reviewed-on: https://chromium-review.googlesource.com/833749
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#525102}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1c36bfd4b104d7f3899942f0e183720a0f011980