Unify the handling of is_issued_by_known_root across platforms
Previously, is_issued_by_known_root took a per-platform approach,
attempting to determine whether or not (for given the platform),
a certificate was issued by a known root. Programmatic policies
such as HPKP or enforcement of the Baseline Requirements was then
gated on this functionality, and reporting errors or additional
details was similarly limited to that set of known roots.

However, this results in inconsistency across platforms - a given
certificate may be a known root on Windows, while returning false
on other platforms, giving an inconsistent security experience and
inconsistent user experience, even though the set of expectations
for that root cert/CA remain the same.

In order to align behaviour, all platforms now consider the set of
unified trust anchors for all platforms, in addition to the runtime
determination (when available) to determine the status. This will
reduce the false-negative case across platforms.

To avoid situations in which the static/compiled in list drifts
from the OS (e.g. due to emergency updates or pre-release versions),
the OS will still be consulted, but it will be as a fallback to
the internal list.

BUG=798186
TBR=isherman@chromium.org

Change-Id: I46e7abf044682b3583bd3ea50369891f5c56c0a6
Reviewed-on: https://chromium-review.googlesource.com/854014
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#531377}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36f20e46515ab61df4ae4af9655b647bf9a0ea5a