Retag idle SpdySessions when streams are requested with different socket tags
This change should alleviate a problem with socket tagging and HTTP/2:
HTTP/2 sessions (SpdySessions) when idle still continue to occupy a socket
pool slot which counts towards the 6-connections-per-host limit. This means
that when six requests with different socket tags are made to the same host,
they will max out the connections-per-host allotment, and a seventh request
will block until one of the HTTP/2 sessions hits the idle timeout and closes.
This isn't a problem for HTTP/1 because we can retag an idle socket and
reuse it to service a new request with a different socket tag.
This isn't a problem for QUIC because there isn't such a
6-connections-per-host limit.

This change allows an idle SpdySession to be retagged with a new socket tag
so that it can be used to service a new request. The following things
happen during this retagging:
1. The underlying socket is tagged with the new socket tag.
2. The SpdySession's SpdySessionKey's socket tag is updated to the new tag.
3. The SpdySessionPool's pool maps are updated to the new session's tag.

Pointers to SpdySessions are handed out to SpdyHttpStreams in advance of the
SpdyHttpStream initiating a SpdyStreamRequest.  The SpdySession could have
its socket tag changed during this interval between when the SpdySession
pointer is handed out to the SpdyHttpStream and when the SpdyHttpStream
initiates the SpdyStreamRequest.  To avoid SpdyStreamRequests being
initiated on SpdySessions with changed tags, this change includes the
desired tag in the SpdyStreamRequest.StartRequest() call arguments so it
can be checked against the SpdySession's current tag.  In reality the
chance of this mismatch is negligible because SpdyStreamRequests are
initiated immediately after the SpdySession is created (see how
HttpNetworkTransaction calls InitializeStream() immediately after creating
the stream), but it's still mitigated. Once InitializeStream() is called,
the SpdySession will no longer be considered idle, so its socket tag
cannot be changed.

Bug: 520198
Change-Id: I6bfc4a9d0802dda924c13806a6d863cfb1e2b3f9
Reviewed-on: https://chromium-review.googlesource.com/938772
Commit-Queue: Paul Jensen <pauljensen@chromium.org>
Reviewed-by: Bence Béky <bnc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#540052}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 94cde5f153c82ee9bafd74bde34246fc88230701