QUIC connections will always have at least one retransmittable packet on the wire if |retransmittable_on_wire_timeout_| is set and there are active requests.
If retransmittable_on_wire_timeout_ is explicitly set, then the following behavior is enabled:
If an ACK is received such that there are no more unacked packets, a short delay timer will start if the connection has active requests (to expire retransmittable_on_wire_timeout_ from the now). When the timer expires, a PING frame will be sent. If a retransmittable packet is sent while the timer is active, the timer is cancelled.
This is intended to only be enabled for Cronet for connection migration.

Merges internal change 187955841

Change-Id: I0517e2768ccdd785f05de484dc2fe77eaaa32ea0
Reviewed-on: https://chromium-review.googlesource.com/953340
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Commit-Queue: Yixin Wang <wangyix@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#541626}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 32e4b32b7afcac36791ae29af2d5dc7416fba19c