[Http Cache] Do not preserve incomplete compressed responses.
If a response is not complete and has a content-encoding header, it
should not be marked as truncated since we do not want it to lead to
range requests for future requests.

TEST=net_unittests --gtest_filter=
WritersTest.ContentEncodingShouldNotTruncate

Also manually verified that the test case reported in the bug is fixed.

Bug: 820862
Change-Id: Ie3f4098e769e257632058af2215054809ca8c89f
Reviewed-on: https://chromium-review.googlesource.com/962346
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#543481}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1720d2a67b945ba58b008c593f358298b32ff0bb