[Http Cache]: Reset the range headers to original when restarting
HttpCache::Transaction::ResetPartialState dooms the current entry, and
goes back to creating a new entry for this request. In case of parallel
validation, if the first request is a full request and has written some
part of the response to the cache, then this second transaction would
create the range header with as many bytes as the written response even
on dooming the current entry which is incorrect.
This CL fixes it by restoring the original range header on this path.

RangeGET_ParallelValidationCouldntConditionalize
Also manually verified the test case reported in the bug.

TEST: net_unittests --gtest_filter=HttpCache.
Bug: 820599
Change-Id: I4e57003193873e433c3b503b9fc20a843f28c4da
Reviewed-on: https://chromium-review.googlesource.com/959450
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#543521}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4fe20e106bcb130bea7cbffdca6bc6a192e7b1fa