Record when the set of known roots is out of date
Record a histogram in the event the built-in list of known,
publicly-trusted roots is out of date. This happens when the OS
believes a given certificate is a known root, but the built-in
list of SPKIs does not have this root.

BUG=none

Change-Id: I65f2f495f9e936a81e3cbe1b6d6169d4b682e40d
Reviewed-on: https://chromium-review.googlesource.com/1003234
Reviewed-by: David Benjamin <davidben@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#549577}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 251b013b8e986da2f259df505b775633d6e9d233