cachetool: fix get_stream command in batch mode
The strings present in kCommandNames needs to be handled in
ExecuteCommands().

Also fix a random typo.

There is a not-landed patch to use the batch mode:
https://codereview.chromium.org/2112013003/

Due to a bug in this patch the command 3 was never passed to cachetool,
and hence this problem was not noticed.

Bug: None
Change-Id: I0f5dbdf96e0e771b265400f5b87eee356722ccd4
Reviewed-on: https://chromium-review.googlesource.com/1010290
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#550603}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5b9373ea83c09c5b1ef4d3bfbda3835d050b7716