Use CompletionRepeatingCallback for io_callback_.
Many classes in //net have a private io_callback_ member that is
initialized upon class construction and is passed every time an async
operation is called.  This CL introduces CompletionRepeatingCallback and
uses it in seven classes, where this member is used multiple times.
(This part is no-op, since currently base::Callback is aliased to
base::RepeatingCallback, and base::Bind to base::BindRepeating.)

The rationale is that both a CompletionRepeatingCallback and a
CompletionOnceCallback has a scoped_refptr<BindStateBase> deep under the
hood (using a scoped_refptr for OnceCallback seems unnecessary, but this
allows efficient cast from RepeatingCallback to OnceCallback).
Instantiating such a scoped_refptr on the heap is more expensive than
copying an existing CompletionRepeatingCallback (which reuses the
BindStateBase object, only increases its reference count).  Therefore it
is better to create and reuse a CompletionRepeatingCallback then to bind
a new CompletionOnceCallback every time in classes that use io_callback_
often.

On the other hand, HttpStreamFactoryImpl::JobController::io_callback_ is
only used once, for proxy resolution, and only if there is a proxy
involved.  Therefore it is better to bind this callback at the call
site.  (This CompletionCallback will later be changed to a
CompletionOnceCallback once ProxyResolutionService::ResolveProxy()
supports that.)

Bug: 807724
Change-Id: Ib342b057b8960789514baa21afcda0820a633404
Reviewed-on: https://chromium-review.googlesource.com/1016671
Reviewed-by: Misha Efimov <mef@chromium.org>
Commit-Queue: Bence Béky <bnc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#553546}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 61f756c9b2659a8894615498a787df7b08964327